Исправлены недостатки: - по миганию меню при прорисовке - перенесена проверка правильности даты в класс Date - устранён разный стиль использования библиотек в main. - горизонтальные отступ между самописными библиотеками и системными.

+++ Case.cpp
@@ -31,7 +31,7 @@ Date Case::GetDate()
 \brief метод вывода на экран дела
 \param[in] elLength количество пробелов для выравнивания
 */
-void Case::PrintCase(int elLength)
+void Case::PrintCase(size_t elLength)
 {
 	std::cout << _title;
 	while (elLength-- > 0)
